+++ backend/ambassadors_project/urls.py
@@ -5,7 +5,7 @@ from django.urls import include, path
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
 from rest_framework import routers
 
-from order.urls import router as order_router  # включить в роутер для амбасадора
+from orders.urls import router as orders_router  # включить в роутер для амбасадора
 
 router_v1 = routers.DefaultRouter()
