+++ backend/orders/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class OrderConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "order"
+    name = "orders"
